secret-store: Use protocol attribute for lookup
Without being able to rely on the schema name for the lookup,
we need a different attribute.

This is a hack we can use as long as there is only one protocol (or we would
need to call this multiple times for each protocol, which would also be doable).

With the last commit and without this one we get warnings from chatty:

`WARNING: secret_service_search: must specify at least one attribute to match`