chat-view: Update spinner opacity on history loading change
It seems the spinner may be shown even when ths spinner is not
active.  So when the spinner is not active, also hide it
by setting opacity to zero.  The widget couldn't be hidden
in that case since it will interfere with the hack written
to check if osk is open or not.