chat-view: Simplify updating scrolledwindow view
We are not supposed to change the visiblity of internal children
of any widget. So let's change the scroll policy of scrolled-window.

This is still required because when vscroll policy is set to automatic,
a minimum height is set (for the scroll bar) which is more than
the height we require.