window: Update view and selection only on map
In constructed(), the window object won’t be ready yet.  And if
we try to select a chat, it may result in chatty-application trying
to open the chat.  It will result in g_application_activate() which
will try to create the window again resulting in an infinite loop

Fix it by handling them on map(), so that these actions will be
done only on gtk_window_present().

Fixes #604