mmsd: Don't trust the MMS file type
Chatty takes the mimetype directly from mmsd-tng, which is from the MMS.
The MIME type is just a header in the MMS, and there is no verification
that the file attached is that MIME type in mmsd-tng or in chatty.
As a safety/security precaution, let's look up the MIME type ourselves.