mm-chat: Emit "changed" after unread is completely set
When the signal is emitted, all callback handlers are run immediately.
And so, if the callback sets unread count to zero, it's not actually
saved properly to db as the previous change to db is commited after the
callbacks are run.

Fix it by emitting the signal after all changes are commited

Fixes #689