application: Fix nologin commandline argument
We were loading the manager in startup() which runs before parsing the
commandline arguments, and so the nologin option was never honoured.
So let's delay loading the manager to activate() or in command_line()
in case daemon mode is set which will parse the nologin option before
the manager is asked to load.