ci: Run some tests under asan
We skip the mesage-row test (needs a leak suppression file which is best
done when we build our own image with debug packages for proper
matching) and pgp (as leak fixes are in another file).

Once we have this we can skip the explicit list of tests and use

 --suite chatty

instead (unless we want to run the libcmatrix tests here too).