mm-account: Change SMS validity from 168 minutes to max. allowed value (635040)
Right now the SMS is only valid for 168 minutes (slightly less than 3 hours).
This is not optimal, as the other party may have their phone off and would
never get the SMS. The default is 24 hours, which is slightly better, but
still not optimal. We are changing it to the max value, 635,040 minutes,
or around 63 weeks.