history: Clarify close status
The db handle is used to flag whether the history is still open.

As the db code executes in a thread we need to toggle that flag *past*
calling sqlite's close() as otherwise we might e.g. terminate the
program too early thinking the db is already closed.

In order to avoid confusion rename the the checking function to
chatty_history_is_closed () which only returns TRUE if the database was
actually closed (or never opened).