ma-chat: Don't emit avatar-changed signal for invalid avatar
When no avatar pixbuf can be created for a given GInputStream
the avatar-changed signal should not be emitted.

Otherwise we may end up in a endless loop of trying and
repeatedly failing to get the avatar.

Not emitting the signal breaks the following cycle:

Closes #901

Part-of: <https://gitlab.gnome.org/World/Chatty/-/merge_requests/1381>