treewide: Switch libcmatrix to a subproject
There was a subproject.wrap but it was also registered as a git
submodule. The later makes testing libcmatrix changes harder than needed
(as one needs to add a separate remote in the the subproject, etc).

By just using a subproject wrap we just need to change a single file
(while still having a git repo to do versioned changes in).

The wrap would also allow us to ship patches on top of libcmatrix in
case we e.g. base it on a release and want to apply a targeted fix.

We can also now build a against a "system" libcmatrix which is nice for
e.g. flatpaks.

The downside is that we need to keep the flatpak manifest in sync
manually and that gitlab's tarballs don't contain libcmatrix by
default anymore but rather need to be generated via

     meson dist --include-subprojects …

Part-of: <https://gitlab.gnome.org/World/Chatty/-/merge_requests/1391>