ma-chat: Emit changed when unread count changes
Otherwise the unread counter shows outdated data. Flow so far was as
follows:

- User selects a room
- Chatty invokes set_unread_count
- cm_room_set_read_marker_async updates the read marker
  server side async
- [nothing updates the chat information]

On second click the current unread count is read and the marker
updated.

Now we update the count as soon as we get informed from the server
that the unread count got updated.

Closes: https://gitlab.gnome.org/World/Chatty/-/issues/908
Part-of: <https://gitlab.gnome.org/World/Chatty/-/merge_requests/1392>