message-row: Allow both showing and hiding user details
This change allows messages which had their user details hidden
to show it again later on instead of being hidden forever
(or until the user manages to refresh the chat page some other way).

In the chat page the user details visibility is now always set
to ensure that the visibility is not left in a stale state.

Closes: https://gitlab.gnome.org/World/Chatty/-/issues/912
Part-of: <https://gitlab.gnome.org/World/Chatty/-/merge_requests/1423>