attachment: Return early on task cancellation
The cancellation check should occur as early as possible
to avoid accessing invalid memory.
In particular don't try accessing `self`
as it may be getting finalized.

Note: This is not a problem currently
as all async operations are holding a reference.

Signed-off-by: Evangelos Ribeiro Tzaras <devrtz@fortysixandtwo.eu>
Part-of: <https://gitlab.gnome.org/World/Chatty/-/merge_requests/1448>