attachment: Get rid of AttachmentData struct
If we move the GtkImage into the ChattyAttachment itself,
we can get rid of the AttachmentData struct.

While we're at it, don't take a ref for the callback user data.
We only cancel the cancellable in finalize() and taking a ref
prevents the cancellation from ever getting triggered.

Signed-off-by: Evangelos Ribeiro Tzaras <devrtz@fortysixandtwo.eu>
Part-of: <https://gitlab.gnome.org/World/Chatty/-/merge_requests/1448>