chatty-settings-dialog: Remove explicit finalization of adw dialog
Deleting an existing account causes a segmentation fault, resulting in a crash and multiple critical warnings.

The AdwAlertDialog object is implicitly finalized at the end of its lifecycle,
so there is no need to explicitly free it.
Removing this explicit finalization prevents the app from crashing when deleting an account.

Fixes: https://gitlab.gnome.org/World/Chatty/-/issues/920
Fixes: eccb92d4 ("dialogs/chatty-settings-dialog.c: update error dialog to GTK 4.10")
Part-of: <https://gitlab.gnome.org/World/Chatty/-/merge_requests/1428>