notification: Let the notification daemon handle feedback
Chatty emitting feedback without bringing a window to the foreground (as
e.g. calls does on incoming calls) is very confusing when the user e.g.
disables notifications as the device keeps buzzing without any visual
indication.

It would make sense to emit feedback when chatty is in the foreground
and rather not put out a notification but let's fix that in a different
step (see #943).

Closes: https://gitlab.gnome.org/World/Chatty/-/issues/657
Part-of: <https://gitlab.gnome.org/World/Chatty/-/merge_requests/1431>