window: Ensure activating search allows user input
Activating the search action (through the acccelerator binding)
is not very useful when the view is collapsed to show the chat page:
The chat page will stay visible and the back button in the header
will get focused.

Ensure we switch to the sidebar when collapsed,
so that the user may perform the search.

Signed-off-by: Evangelos Ribeiro Tzaras <devrtz@fortysixandtwo.eu>
Part-of: <https://gitlab.gnome.org/World/Chatty/-/merge_requests/1451>