side-bar: Only toggle search off if focussed
And grab focus first otherwise.

When the search is toggled on, but the focus is somewhere else,
e.g. in the chat page, a user would have to press ctrl+f twice
to get to the search.

The first accelerator activation will toggle the search off
(instead of focusing the entry)
and the second activation will toggle the search on
including focusing the entry, allowing direct user input.

Signed-off-by: Evangelos Ribeiro Tzaras <devrtz@fortysixandtwo.eu>
Part-of: <https://gitlab.gnome.org/World/Chatty/-/merge_requests/1451>