mms: ensure we don't squash attachments with the same name
Some devices strip out the name of their attachments and replace
them with a generic string (like "image"). That means that if we get
two images from a device like that, we will squash the first file with
the second one, which will result in the user seeing the same image
twice. So now we prepend a monotonically increasing number to the
filename to ensure that conflicts like that can't ever happen.